Implement SHA-3 (#17)
* Implement SHA-3

This implementation is a redesign of the version in libkeccak, with a focus
on small code size by avoiding code bloat from too many different generic
instantiations and avoiding unnecessary features. In particular, this
implementation supports only bytewise messages (i.e. the length of all
inputs and outputs must be a multiple of 8 bits) and the Keccak permutation
only supports XORing and reading complete lanes at a time.

One implementation of Keccak is provided, which is the "lane complemented"
implementation. This implementation reduces the number of NOT operations to 1
per plane by storing the complement of certain lanes.

* Add SHA-3 test vectors

* Add benchmarks for Keccak-p[1600, 24] and Keccak-p[1600, 12]

* Add SHA-3 unit tests

* Add SHA-3 configuration variables to CI workflow

* Add SHA-3 self-tests

* Add SHA-3 to the list of hash functions in the README