Implement SHAKE
This implements the two extendable output functions (XOF) defined
in NIST FIPS PUB 202: SHAKE128 and SHAKE256.

A new convenience package, Tux.XOF, is added which fulfils the same
role as Tux.Hashing, but for XOFs.

The Keccak_Enabled crate configuration variable has been removed as
this is not really necessary. Calls to Keccak will be removed when
SHA-3 and SHAKE are disabled in the crate configuration, and so the
Keccak code will be automatically stripped by the linker (assuming that
there are no other calls to Keccak in user code).