Add SHAKE test vectors
This adds the NIST test vectors for SHAKE128 and SHAKE256.

Two new KAT test driver programs are added:
 * 'xof' does the same as the 'hash' test driver, but for XOFs.
 * 'xof_monte' runs the Monte Carlo pseudorandom tests for XOFs
   and generates all 100 checkpoint values.

Note that for the Monte Carlo tests a different testing approach is
needed than compared to the hash tests since the output length is
variable and needs to be tracked throughout all 100,000 calculations.