Remove Self_Tests_Enabled crate configuration variable.
This configuration variable adds extra complexity and is actually unnecessary.
The original intent was to be able to reduce code size by excluding the
self test code from the executable. But explicitly omitting the code is not
necessary since the linker will remove uncalled functions.

When an algorithm is disabled in the crate configuration, the corresponding
self test is replaced with a stub to allow units that "with" the self test
still compile. This also simplifies the test suite since we no longer need
to replace tests with stubs.