Add Keccak backend optimized for speed.
The "size" backend is the same as the existing one where a single
"Permute" function is used where the number of rounds are known at
run-time. The "speed" backend uses multiple instantiations of the
"Permute" function where the number of rounds is known statically
to enable optimizations for the specific number of rounds.

This gives about a 0.2 cpb performance improvement in the SHA3-512
benchmark on an Amd Ryzen 5800X CPU with GNAT FSF 12.