Benchmark improvements
Increasing the number of measurements per test makes the results a bit
more stable.

The data sizes that are measured now increase in multiples of 8 up to
256 kiB (down from 1 MiB). This speeds up the benchmark, and measurements
for 256 kiB and above don't really change.

The benchmark now skips measuring algorithms that are disabled in the crate
configuration. This prevents the predicate on Enabled_Algorithm_Type from
being violated.