Keccak performance improvements (#21)
* Add Keccak backend optimized for speed.

The "size" backend is the same as the existing one where a single
"Permute" function is used where the number of rounds are known at
run-time. The "speed" backend uses multiple instantiations of the
"Permute" function where the number of rounds is known statically
to enable optimizations for the specific number of rounds.

This gives about a 0.2 cpb performance improvement in the SHA3-512
benchmark on an Amd Ryzen 5800X CPU with GNAT FSF 12.

* Avoid loop-carried dependencies when xoring/extracting the Keccak state

This reworks the XOR_Bytes_Into_Context and Extract_Bytes procedures to
avoid loop-carried dependencies. This allows Loop_Optimize (Ivdep, Vector)
hints to be used which can help the optimizer.

* Revert change to Keccak context layout

This broke the tests, so revert for now.

* Add generically optimised version of Keccak

This implementation offers better performance than the lane complemented
implementation on targets that have an "and not" instruction.

* Remove remaining remnants of Self_Tests_Eanbled

* Remove temporary compiler flags for testing

* Add Keccak config variables to CI workflow

* Document Keccak config vars in the readme

* Fix incompatible assertion policies

* Fix redundant with clauses and pragmas in Tux.Keccak_1600

* Fix copy/paste errors in the generic optimized version of Keccak

* Improve Loop_Optimization pragmas for Keccak

Benchmarking on x86-64 shows that disabling loop unrolling improves
performance by about 0.2 cpb and has the added benefit of reducing
code size.

* Benchmark improvements

Increasing the number of measurements per test makes the results a bit
more stable.

The data sizes that are measured now increase in multiples of 8 up to
256 kiB (down from 1 MiB). This speeds up the benchmark, and measurements
for 256 kiB and above don't really change.

The benchmark now skips measuring algorithms that are disabled in the crate
configuration. This prevents the predicate on Enabled_Algorithm_Type from
being violated.